requirements.txt
streamlit
pandas
reportlab
bcrypt
psycopg2-binary
streamlit-cookies-controller